# [Dy2stat]Join break cond with while cond in some pattern