<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Others

### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
Others
### Describe
<!-- Describe what this PR does -->
Join break cond with while cond in some pattern.

Consider the follow code:
```python
import paddle
import time

class Net(paddle.nn.Layer):
    def __init__(self):
        super(Net, self).__init__()
        self.linear = paddle.nn.Linear(224, 224)

    @paddle.jit.to_static
    def forward(self, x):
        for i in range(4):
            if x.mean() > 10:
                break
            x = self.linear(x)

        return x
```

Before this PR, `break` statement will introduce a `cond_op` and wraps all behind ops into it. It was proved very time-consuming. The transformed code is :
```python
def forward(self, x):
    __break_0 = paddle.fluid.layers.fill_constant(shape=[1], dtype='bool',
        value=False)
    i = 0

    def while_condition_0(x, __break_0, i):
        return paddle.jit.dy2static.convert_logical_and(x=i + 1 <= 4, y=
            paddle.jit.dy2static.convert_logical_not(__break_0))

    def while_body_0(x, __break_0, i):

        def true_fn_0(__break_0):
            __break_0 = paddle.fluid.layers.fill_constant(shape=[1], dtype=
                'bool', value=True)
            return __break_0

        def false_fn_0(__break_0):
            return __break_0
        __break_0 = paddle.jit.dy2static.convert_ifelse(paddle.jit.
            dy2static.convert_call(x.mean)() > 10, true_fn_0, false_fn_0, (
            __break_0,), (__break_0,), (__break_0,))

        def true_fn_1(x):
            x = paddle.jit.dy2static.convert_call(self.linear)(x)
            return x

        def false_fn_1(x):
            return x
        x = paddle.jit.dy2static.convert_ifelse(paddle.jit.dy2static.
            convert_logical_not(__break_0), true_fn_1, false_fn_1, (x,), (x
            ,), (x,))
        i += 1
        return x, __break_0, i
    [x, __break_0, i] = paddle.jit.dy2static.convert_while_loop(
        while_condition_0, while_body_0, [x, __break_0, i])
    return x
``` 

In this PR, we join break conditional statement with while to avoid the overhead of`cond_op. 
The transformed code is:
```python
def forward(self, x):
    i = 0

    def while_condition_0(x, i):
        return paddle.jit.dy2static.convert_logical_and(x=i + 1 <= 4, y=
            paddle.jit.dy2static.convert_logical_not(paddle.jit.dy2static.
            convert_call(x.mean)() > 10))

    def while_body_0(x, i):
        x = paddle.jit.dy2static.convert_call(self.linear)(x)
        i += 1
        return x, i
    [x, i] = paddle.jit.dy2static.convert_while_loop(while_condition_0,
        while_body_0, [x, i])
    return x
```

We evaluate the cost time by:
```python

net = Net()

for i in range(10):
    start = time.time()
    out = net(x)
    end = time.time()
    print('cost time: ', end - start)
```

| Before(static)| After(static) |Dygraph |Speed up|
|:----:|:----:|:----:|:----:|
|85 ms| 11 ms |9 ms|7x|