@Aurelius84 2020-10-22
> Is it easy to split optimization transform and break_continue_transformer?
> 
> I think an easy-to-manage organization in the future is that we keep the old functional transformer classes and have some optimization transformer classes (Or to make it distinguishable, we can rename "transformer" for optimization to something else)

That's cool. I will push a new commit to split as optimization.